llm/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
torch==2.4.0
transformers==4.43.3
